requirements.txt
# Next Up: NBA G-League Call-Up Prediction
# Python dependencies for data collection, analysis, and modeling

# Core data processing
pandas>=1.5.0
numpy>=1.23.0

# API and web requests
requests>=2.28.0
python-dotenv>=0.21.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Machine learning
scikit-learn>=1.2.0
xgboost>=1.7.0

# Model interpretation
shap>=0.41.0

# Web app framework
streamlit>=1.20.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.20.0

# Optional: Web scraping (if SportsRadar API unavailable)
beautifulsoup4>=4.11.0
selenium>=4.7.0
lxml>=4.9.0

# Optional: Advanced features
imbalanced-learn>=0.10.0  # For handling class imbalance
optuna>=3.1.0  # For hyperparameter tuning

